% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhmkin.R
\name{check_failed}
\alias{check_failed}
\title{Check if fit within an mhmkin object failed}
\usage{
check_failed(x)
}
\arguments{
\item{x}{The object to be checked}
}
\description{
Check if fit within an mhmkin object failed
}
