% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\alias{convergence.mmkin}
\alias{print.convergence.mmkin}
\title{Method to get convergence information}
\usage{
convergence(object, ...)

\method{convergence}{mmkin}(object, ...)

\method{print}{convergence.mmkin}(x, ...)
}
\arguments{
\item{object}{The object to investigate}

\item{\dots}{For potential future extensions}

\item{x}{The object to be printed}
}
\value{
For \link{mkinfit} objects, a character vector containing
For \link{mmkin} objects, an object of class 'convergence.mmkin' with a
suitable printing method.
}
\description{
Method to get convergence information
}
\examples{
\dontrun{
fits <- mmkin(
  c("SFO", "FOMC"),
  list("FOCUS A" = FOCUS_2006_A,
       "FOCUS B" = FOCUS_2006_C),
  quiet = TRUE)
convergence(fits)
}
}
