% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimethenamid_2018.R
\docType{data}
\name{dimethenamid_2018}
\alias{dimethenamid_2018}
\title{Aerobic soil degradation data on dimethenamid and dimethenamid-P from the EU assessment in 2018}
\format{
An \link{mkindsg} object grouping seven datasets with some meta information
}
\source{
Rapporteur Member State Germany, Co-Rapporteur Member State Bulgaria (2018)
Renewal Assessment Report Dimethenamid-P Volume 3 - B.8 Environmental fate and behaviour
Rev. 2 - November 2017
\href{https://open.efsa.europa.eu}{https://open.efsa.europa.eu/study-inventory/EFSA-Q-2014-00716}
}
\usage{
dimethenamid_2018
}
\description{
The datasets were extracted from the active substance evaluation dossier
published by EFSA. Kinetic evaluations shown for these datasets are intended
to illustrate and advance kinetic modelling. The fact that these data and
some results are shown here does not imply a license to use them in the
context of pesticide  registrations, as the use of the data may be
constrained by data protection regulations.
}
\details{
The R code used to create this data object is installed with this package
in the 'dataset_generation' directory. In the code, page numbers are given for
specific pieces of information in the comments.
}
\examples{
print(dimethenamid_2018)
}
\keyword{datasets}
