% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_mixed.R
\name{ds_mixed}
\alias{ds_mixed}
\alias{ds_sfo}
\alias{ds_fomc}
\alias{ds_dfop}
\alias{ds_hs}
\alias{ds_dfop_sfo}
\title{Synthetic data for hierarchical kinetic degradation models}
\description{
The R code used to create this data object is installed with this package in
the 'dataset_generation' directory.
}
\examples{
\dontrun{
  sfo_mmkin <- mmkin("SFO", ds_sfo, quiet = TRUE, error_model = "tc", cores = 15)
  sfo_saem <- saem(sfo_mmkin, no_random_effect = "parent_0")
  plot(sfo_saem)
}

# This is the code used to generate the datasets
cat(readLines(system.file("dataset_generation/ds_mixed.R", package = "mkin")), sep = "\n")
}
