\name{FOCUS_2006_datasets}
\alias{FOCUS_2006_A}
\alias{FOCUS_2006_B}
\alias{FOCUS_2006_C}
\alias{FOCUS_2006_D}
\alias{FOCUS_2006_E}
\alias{FOCUS_2006_F}
\docType{data}
\title{
Datasets A to F from the FOCUS Kinetics report from 2006
}
\description{
Data taken from FOCUS (2006), p. 258.
}
\usage{FOCUS_2006_datasets}
\format{
  6 datasets with observations on the following variables.
  \describe{
    \item{\code{name}}{a factor containing the name of the observed variable}
    \item{\code{time}}{a numeric vector containing time points}
    \item{\code{value}}{a numeric vector containing concentrations in percent of applied radioactivity}
  }
}
\source{
  FOCUS (2006) \dQuote{Guidance Document on Estimating Persistence and
  Degradation Kinetics from Environmental Fate Studies on Pesticides in EU
  Registration} Report of the FOCUS Work Group on Degradation Kinetics,
  EC Document Reference Sanco/10058/2005 version 2.0, 434 pp,
  \url{http://focus.jrc.ec.europa.eu/dk} 
}
\examples{
FOCUS_2006_C
}
\keyword{datasets}
