% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_soil_moisture.R
\docType{data}
\name{focus_soil_moisture}
\alias{focus_soil_moisture}
\title{FOCUS default values for soil moisture contents at field capacity, MWHC and 1/3 bar}
\format{
A matrix with upper case USDA soil classes as row names, and water tension
('pF1', 'pF2', 'pF 2.5') as column names
}
\source{
Anonymous (2014) Generic Guidance for Tier 1 FOCUS Ground Water Assessment
Version 2.2, May 2014 \url{https://esdac.jrc.ec.europa.eu/projects/ground-water}
}
\usage{
focus_soil_moisture
}
\description{
FOCUS default values for soil moisture contents at field capacity, MWHC and 1/3 bar
}
\examples{
focus_soil_moisture
}
\keyword{datasets}
