% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme.mmkin.R
\name{get_deg_func}
\alias{get_deg_func}
\title{Retrieve a degradation function from the mmkin namespace}
\usage{
get_deg_func()
}
\value{
A function that was likely previously assigned from within
  nlme.mmkin
}
\description{
Retrieve a degradation function from the mmkin namespace
}
