\name{gmkin}
\alias{gmkin}
\title{
  Start a graphical user interface (GUI) based on the \code{gWidgetsWWW2} toolkit.
}
\description{
  This function starts a browser based GUI. Please visit the 
  \href{http://github.com/jverzani/gWidgetsWWW2}{github page of gWidgetsWWW2} 
  for an explanation how this toolkit works.
}
\usage{
  gmkin()
}
\value{
  The function is called for its side effect, namely starting the GUI in a browser.
}
\author{
  Johannes Ranke
}
