% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_kinetics.R
\name{hierarchical_kinetics}
\alias{hierarchical_kinetics}
\title{Hierarchical kinetics template}
\usage{
hierarchical_kinetics(..., keep_tex = FALSE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
\code{\link[rmarkdown:render]{render}}
}
\description{
R markdown format for setting up hierarchical kinetics based on a template
provided with the mkin package.
}
\examples{

\dontrun{
library(rmarkdown)
draft("New analysis.rmd", template = "hierarchical_kinetics", package = "mkin")
}

}
