% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals.nlmixr.mmkin}
\alias{intervals.nlmixr.mmkin}
\title{Confidence intervals for parameters in nlmixr.mmkin objects}
\usage{
\method{intervals}{nlmixr.mmkin}(object, level = 0.95, backtransform = TRUE, ...)
}
\arguments{
\item{object}{The fitted saem.mmkin object}

\item{level}{The confidence level.}

\item{backtransform}{Should we backtransform the parameters where a one to
one correlation between transformed and backtransformed parameters exists?}

\item{\dots}{For compatibility with the generic method}
}
\value{
An object with 'intervals.saem.mmkin' and 'intervals.lme' in the
class attribute
}
\description{
Confidence intervals for parameters in nlmixr.mmkin objects
}
