% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals.saem.mmkin}
\alias{intervals.saem.mmkin}
\title{Confidence intervals for parameters in saem.mmkin objects}
\usage{
\method{intervals}{saem.mmkin}(object, level = 0.95, backtransform = TRUE, ...)
}
\arguments{
\item{object}{The fitted saem.mmkin object}

\item{level}{The confidence level. Must be the default of 0.95 as this is what
is available in the saemix object}

\item{backtransform}{In case the model was fitted with mkin transformations,
should we backtransform the parameters where a one to one correlation
between transformed and backtransformed parameters exists?}

\item{\dots}{For compatibility with the generic method}
}
\value{
An object with 'intervals.saem.mmkin' and 'intervals.lme' in the
class attribute
}
\description{
Confidence intervals for parameters in saem.mmkin objects
}
