% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llhist.R
\name{llhist}
\alias{llhist}
\title{Plot the distribution of log likelihoods from multistart objects}
\usage{
llhist(object, breaks = "Sturges", lpos = "topleft", main = "", ...)
}
\arguments{
\item{object}{The \link{multistart} object}

\item{breaks}{Passed to \link{hist}}

\item{lpos}{Positioning of the legend.}

\item{main}{Title of the plot}

\item{\dots}{Passed to \link{hist}}
}
\description{
Produces a histogram of log-likelihoods, and an overlayed kernel density
estimate. In addition, the likelihood of the original fit is shown as
a red vertical line.
}
\seealso{
\link{multistart}
}
