% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem.R
\name{logLik.saem.mmkin}
\alias{logLik.saem.mmkin}
\title{logLik method for saem.mmkin objects}
\usage{
\method{logLik}{saem.mmkin}(object, ..., method = c("is", "lin", "gq"))
}
\arguments{
\item{object}{The fitted \link{saem.mmkin} object}

\item{\dots}{Passed to \link[saemix:logLik]{saemix::logLik.SaemixObject}}

\item{method}{Passed to \link[saemix:logLik]{saemix::logLik.SaemixObject}}
}
\description{
logLik method for saem.mmkin objects
}
