% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mkinfit.R
\name{mkinplot}
\alias{mkinplot}
\title{Plot the observed data and the fitted model of an mkinfit object}
\usage{
mkinplot(fit, ...)
}
\arguments{
\item{fit}{an object of class \code{\link{mkinfit}}.}

\item{\dots}{further arguments passed to \code{\link{plot.mkinfit}}.}
}
\value{
The function is called for its side effect.
}
\description{
Deprecated function. It now only calls the plot method
\code{\link{plot.mkinfit}}.
}
\author{
Johannes Ranke
}
