% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.mkinfit.R
\name{nobs.mkinfit}
\alias{nobs.mkinfit}
\title{Number of observations on which an mkinfit object was fitted}
\usage{
\method{nobs}{mkinfit}(object, ...)
}
\arguments{
\item{object}{An mkinfit object}

\item{\dots}{For compatibility with the generic method}
}
\value{
The number of rows in the data included in the mkinfit object
}
\description{
Number of observations on which an mkinfit object was fitted
}
