% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parhist.R
\name{parhist}
\alias{parhist}
\title{Plot parameter distributions from multistart objects}
\usage{
parhist(object, lpos = "topleft", main = "", ...)
}
\arguments{
\item{object}{The \link{multistart} object}

\item{lpos}{Positioning of the legend.}

\item{main}{Title of the plot}

\item{\dots}{Passed to \link{boxplot}}
}
\description{
Produces a boxplot with all parameters from the multiple runs, scaled using
their medians. If parameter transformations were done by mkin (default in
\link{saem}), then the parameters found by saem are on the transformed scale, and
scaling is simply done by subtracting the median. If parameter
transformations were done by saemix, scaling is done by dividing by the
median, as in the paper by Duchesne et al. (2021).
}
\references{
Duchesne R, Guillemin A, Gandrillon O, Crauste F. Practical
identifiability in the frame of nonlinear mixed effects models: the example
of the in vitro erythropoiesis. BMC Bioinformatics. 2021 Oct 4;22(1):478.
doi: 10.1186/s12859-021-04373-4.
}
\seealso{
\link{multistart}
}
