% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkinds.R
\name{print.mkinds}
\alias{print.mkinds}
\title{Print mkinds objects}
\usage{
\method{print}{mkinds}(x, ...)
}
\arguments{
\item{x}{An \code{\link{mkinds}} object.}

\item{\dots}{Not used.}
}
\description{
Print mkinds objects.
}
