% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmkin.R
\name{print.mmkin}
\alias{print.mmkin}
\title{Print method for mmkin objects}
\usage{
\method{print}{mmkin}(x, ...)
}
\arguments{
\item{x}{An \link{mmkin} object.}

\item{\dots}{Not used.}
}
\description{
Print method for mmkin objects
}
