% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.mkinfit.R
\name{residuals.mkinfit}
\alias{residuals.mkinfit}
\title{Extract residuals from an mkinfit model}
\usage{
\method{residuals}{mkinfit}(object, standardized = FALSE, ...)
}
\arguments{
\item{object}{An \code{\link{mkinfit}} object}

\item{standardized}{Should the residuals be standardized by dividing by the
standard deviation obtained from the fitted error model?}

\item{\dots}{Not used}
}
\description{
Extract residuals from an mkinfit model
}
\examples{
f <- mkinfit("DFOP", FOCUS_2006_C, quiet = TRUE)
residuals(f)
residuals(f, standardized = TRUE)
}
