% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\alias{status.mmkin}
\alias{print.status.mmkin}
\title{Method to get status information for fit array objects}
\usage{
status(object, ...)

\method{status}{mmkin}(object, ...)

\method{print}{status.mmkin}(x, ...)
}
\arguments{
\item{object}{The object to investigate}

\item{\dots}{For potential future extensions}

\item{x}{The object to be printed}
}
\value{
An object with the same dimensions as the fit array
suitable printing method.
}
\description{
Method to get status information for fit array objects
}
\examples{
\dontrun{
fits <- mmkin(
  c("SFO", "FOMC"),
  list("FOCUS A" = FOCUS_2006_A,
       "FOCUS B" = FOCUS_2006_C),
  quiet = TRUE)
status(fits)
}
}
