% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_listing.R
\name{summary_listing}
\alias{summary_listing}
\alias{tex_listing}
\alias{html_listing}
\title{Display the output of a summary function according to the output format}
\usage{
summary_listing(object, caption = NULL, label = NULL, clearpage = TRUE)

tex_listing(object, caption = NULL, label = NULL, clearpage = TRUE)

html_listing(object, caption = NULL)
}
\arguments{
\item{object}{The object for which the summary is to be listed}

\item{caption}{An optional caption}

\item{label}{An optional label, ignored in html output}

\item{clearpage}{Should a new page be started after the listing? Ignored in html output}
}
\description{
This function is intended for use in a R markdown code chunk with the chunk
option \code{results = "asis"}.
}
