% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_listing.R
\name{tex_listing}
\alias{tex_listing}
\title{Wrap the output of a summary function in tex listing environment}
\usage{
tex_listing(object, caption = NULL, label = NULL, clearpage = TRUE)
}
\arguments{
\item{object}{The object for which the summary is to be listed}

\item{caption}{An optional caption}

\item{label}{An optional label}

\item{clearpage}{Should a new page be started after the listing?}
}
\description{
Wrap the output of a summary function in tex listing environment
}
