% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tffm0.R
\name{tffm0}
\alias{tffm0}
\alias{invtffm0}
\title{Transform formation fractions as in the first published mkin version}
\usage{
tffm0(ff)

invtffm0(ff_trans)
}
\arguments{
\item{ff}{Vector of untransformed formation fractions. The sum
must be smaller or equal to one}

\item{ff_trans}{Vector of transformed formation fractions that can be
restricted to the interval from 0 to 1}
}
\value{
A vector of the transformed formation fractions

A vector of backtransformed formation fractions for natural use in degradation models
}
\description{
The transformed fractions can be restricted between 0 and 1 in model
optimisations. Therefore this transformation was used originally in mkin. It
was later replaced by the \link{ilr} transformation because the ilr transformed
fractions can assumed to follow normal distribution. As the ilr
transformation is not available in \link{RxODE} and can therefore not be used in
the nlmixr modelling language, this transformation is currently used for
translating mkin models with formation fractions to more than one target
compartment for fitting with nlmixr in \link{nlmixr_model}. However,
this implementation cannot be used there, as it is not accessible
from RxODE.
}
\examples{
ff_example <- c(
  0.10983681, 0.09035905, 0.08399383
)
ff_example_trans <- tffm0(ff_example)
invtffm0(ff_example_trans)
}
