% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.mkinfit.R
\name{confint.mkinfit}
\alias{confint.mkinfit}
\title{Confidence intervals for parameters of mkinfit objects}
\usage{
\method{confint}{mkinfit}(object, parm, level = 0.95, alpha = 1 -
  level, method = c("profile", "quadratic"), transformed = TRUE,
  backtransform = TRUE, distribution = c("student_t", "normal"),
  quiet = FALSE, ...)
}
\arguments{
\item{object}{An \code{\link{mkinfit}} object}

\item{parm}{A vector of names of the parameters which are to be given
confidence intervals. If missing, all parameters are considered.}

\item{level}{The confidence level required}

\item{alpha}{The allowed error probability, overrides 'level' if specified.}

\item{method}{The 'profile' method searches the parameter space for the
cutoff of the confidence intervals by means of a likelihood ratio test.
The 'quadratic' method approximates the likelihood function at the
optimised parameters using the second term of the Taylor expansion, using
a second derivative (hessian) contained in the object.}

\item{transformed}{If the quadratic approximation is used, should it be
applied to the likelihood based on the transformed parameters?}

\item{backtransform}{If we approximate the likelihood in terms of the
transformed parameters, should we backtransform the parameters with
their confidence intervals?}

\item{distribution}{For the quadratic approximation, should we use
the student t distribution or assume normal distribution for
the parameter estimate}

\item{quiet}{Should we suppress messages?}
}
\value{
A matrix with columns giving lower and upper confidence limits for
  each parameter.
}
\description{
Confidence intervals for parameters of mkinfit objects
}
\examples{
f <- mkinfit("SFO", FOCUS_2006_C, quiet = TRUE)
confint(f, method = "quadratic")
confint(f, method = "profile")
}
\references{
Pawitan Y (2013) In all likelihood - Statistical modelling and
  inference using likelihood. Clarendon Press, Oxford.
}
