% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deg_func.R
\name{create_deg_func}
\alias{create_deg_func}
\title{Create degradation functions for known analytical solutions}
\usage{
create_deg_func(spec, use_of_ff = c("min", "max"))
}
\arguments{
\item{spec}{List of model specifications as contained in mkinmod objects}

\item{use_of_ff}{Minimum or maximum use of formation fractions}
}
\value{
Degradation function to be attached to mkinmod objects
}
\description{
Create degradation functions for known analytical solutions
}
\examples{

SFO_SFO <- mkinmod(
  parent = mkinsub("SFO", "m1"),
  m1 = mkinsub("SFO"))
fit <- mkinfit(SFO_SFO, FOCUS_2006_D, quiet = TRUE)
}
