% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_solutions.R
\name{HS.solution}
\alias{HS.solution}
\title{Hockey-Stick kinetics}
\usage{
HS.solution(t, parent_0, k1, k2, tb)
}
\arguments{
\item{tb}{Break point. Before this time, exponential decline according to
\code{k1} is calculated, after this time, exponential decline proceeds
according to \code{k2}.}
}
\description{
Function describing two exponential decline functions with a break point
between them.
}
\examples{

  plot(function(x) HS.solution(x, 100, 2, 0.3, 0.5), 0, 2, ylim=c(0,100))

}
\seealso{
Other parent solutions: 
\code{\link{DFOP.solution}()},
\code{\link{FOMC.solution}()},
\code{\link{IORE.solution}()},
\code{\link{SFO.solution}()},
\code{\link{SFORB.solution}()},
\code{\link{logistic.solution}()}
}
\concept{parent solutions}
