% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illparms.R
\name{illparms}
\alias{illparms}
\alias{illparms.mkinfit}
\alias{illparms.mmkin}
\alias{print.illparms.mmkin}
\title{Method to get the names of ill-defined parameters}
\usage{
illparms(object, ...)

\method{illparms}{mkinfit}(object, conf.level = 0.95, ...)

\method{illparms}{mmkin}(object, conf.level = 0.95, ...)

\method{print}{illparms.mmkin}(x, ...)
}
\arguments{
\item{object}{The object to investigate}

\item{\dots}{For potential future extensions}

\item{conf.level}{The confidence level for checking p values}

\item{x}{The object to be printed}
}
\value{
For \link{mkinfit} objects, a character vector of parameter names
For \link{mmkin} objects, an object of class 'illparms.mmkin' with a
suitable printing method.
}
\description{
Method to get the names of ill-defined parameters
}
\examples{
fit <- mkinfit("FOMC", FOCUS_2006_A, quiet = TRUE)
illparms(fit)
\dontrun{
fits <- mmkin(
  c("SFO", "FOMC"),
  list("FOCUS A" = FOCUS_2006_A,
       "FOCUS C" = FOCUS_2006_C),
  quiet = TRUE)
illparms(fits)
}
}
