% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saem.R
\name{logLik.saem.mmkin}
\alias{logLik.saem.mmkin}
\title{logLik method for saem.mmkin objects}
\usage{
\method{logLik}{saem.mmkin}(object, ..., method = c("lin", "is", "gq"))
}
\arguments{
\item{method}{Passed to \link[saemix:logLik]{saemix::logLik.SaemixObject}}
}
\description{
logLik method for saem.mmkin objects
}
