% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkinds.R
\docType{class}
\name{mkinds}
\alias{mkinds}
\title{A dataset class for mkin}
\format{An \code{\link{R6Class}} generator object.}
\usage{
mkinds
}
\description{
A dataset class for mkin
}
\section{Fields}{


\describe{ \item{list("title")}{A full title for the dataset}

\item{list("sampling")}{times The sampling times}

\item{list("time_unit")}{The time unit}

\item{list("observed")}{Names of the observed compounds}

\item{list("unit")}{The unit of the observations}

\item{list("replicates")}{The number of replicates}

\item{list("data")}{A dataframe with at least the columns name, time and
value in order to be compatible with mkinfit} }
}

\examples{

mds <- mkinds$new("FOCUS A", FOCUS_2006_A)

}
\keyword{datasets}
