\name{mkinplot}
\alias{mkinplot}
\title{
  Plot the observed data and the fitted model of an mkinfit.
}
\description{
  Deprecated function. It now only calls the plot method \code{\link{plot.mkinfit}}.
}
\usage{
  mkinplot(fit, ...)
}
\arguments{
  \item{fit}{
    an object of class \code{\link{mkinfit}}.
  }
  \item{\dots}{
   further arguments passed to \code{\link{plot.mkinfit}}.
}
}
\value{
  The function is called for its side effect.
}
\author{
  Johannes Ranke
}
\keyword{ hplot }
