% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart.R
\name{multistart}
\alias{multistart}
\alias{multistart.saem.mmkin}
\alias{print.multistart}
\alias{parms.multistart}
\alias{parhist}
\alias{llhist}
\title{Perform a hierarchical model fit with multiple starting values}
\usage{
multistart(object, n = 50, cores = 1, ...)

\method{multistart}{saem.mmkin}(object, n = 50, cores = 1, ...)

\method{print}{multistart}(x, ...)

\method{parms}{multistart}(object, ...)

parhist(object, lpos = "topleft", main = "", ...)

llhist(object, breaks = "Sturges", lpos = "topleft", main = "", ...)
}
\arguments{
\item{object}{The fit object to work with}

\item{n}{How many different combinations of starting parameters should be
used?}

\item{cores}{How many fits should be run in parallel?}

\item{\dots}{Passed to the update function, or to the basic plotting
function in the case of the graphical functions.}

\item{x}{The multistart object to print}

\item{lpos}{Positioning of the legend.}

\item{main}{title of the plot}

\item{breaks}{Passed to \link{hist}}
}
\value{
A list of \link{saem.mmkin} objects, with class attributes
'multistart.saem.mmkin' and 'multistart'.
}
\description{
The purpose of this method is to check if a certain algorithm for fitting
nonlinear hierarchical models (also known as nonlinear mixed-effects models)
will reliably yield results that are sufficiently similar to each other, if
started with a certain range of reasonable starting parameters. It is
inspired by the article on practical identifiabiliy in the frame of nonlinear
mixed-effects models by Duchesne et al (2021).
}
\details{
Currently, parallel execution of the fits is only supported using
\link[parallel:mclapply]{parallel::mclapply}, i.e. not available on Windows.
}
\references{
Duchesne R, Guillemin A, Gandrillon O, Crauste F. Practical
identifiability in the frame of nonlinear mixed effects models: the example
of the in vitro erythropoiesis. BMC Bioinformatics. 2021 Oct 4;22(1):478.
doi: 10.1186/s12859-021-04373-4.
}
