\name{print.mkinmod}
\alias{print.mkinmod}
\title{
  Print mkinmod objects
}
\description{
  Print mkinmod objects in a way that the user finds his way to get to its components.
}
\usage{
\method{print}{mkinmod}(x, ...)
}
\arguments{
  \item{x}{
    An \code{\link{mkinmod}} object.
  }
  \item{\dots}{
    Not used.
  }
}
