\name{print.nafta}
\alias{print.nafta}
\title{
  Print nafta objects
}
\description{
  Print nafta objects.
}
\usage{
  \method{print}{nafta}(x, quiet = TRUE, ...)
}
\arguments{
  \item{x}{
    An \code{\link{nafta}} object.
  }
  \item{quiet}{
    Should the evaluation text be shown?
  }
  \item{\dots}{
    Not used.
  }
}
