% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.mkinfit.R, R/nlme.mmkin.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lrtest}
\alias{nlme}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lmtest}{\code{\link[lmtest]{lrtest}}}

  \item{nlme}{\code{\link[nlme]{nlme}}}
}}

