% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saemix.R
\name{saemix_model}
\alias{saemix_model}
\alias{saemix_data}
\title{Create saemix models from mmkin row objects}
\usage{
saemix_model(object)

saemix_data(object, ...)
}
\arguments{
\item{object}{An mmkin row object containing several fits of the same model to different datasets}

\item{\dots}{Further parameters passed to \link[saemix:saemixData]{saemix::saemixData}}
}
\value{
An \link[saemix:SaemixModel]{saemix::SaemixModel} object.

An \link[saemix:SaemixData]{saemix::SaemixData} object.
}
\description{
This function sets up a nonlinear mixed effects model for an mmkin row
object for use with the saemix package. An mmkin row object is essentially a
list of mkinfit objects that have been obtained by fitting the same model to
a list of datasets.
}
\examples{
ds <- lapply(experimental_data_for_UBA_2019[6:10],
 function(x) subset(x$data[c("name", "time", "value")]))
names(ds) <- paste("Dataset", 6:10)
sfo_sfo <- mkinmod(parent = mkinsub("SFO", "A1"),
  A1 = mkinsub("SFO"))
f_mmkin <- mmkin(list("SFO-SFO" = sfo_sfo), ds, quiet = TRUE, cores = 5)
\dontrun{
if (require(saemix)) {
  m_saemix <- saemix_model(f_mmkin)
  d_saemix <- saemix_data(f_mmkin)
  saemix_options <- list(seed = 123456, save = FALSE, save.graphs = FALSE)
    saemix(m_saemix, d_saemix, saemix_options)
}
}
}
